Fix: Update end selection mark in removing surrounding text
I submitted a pull request #3, which included the commit
2b73508c9becd962da9c2c7ca38b3ac32ebd27e1, to address the reselection of
text using the 'gv' command. In my previous implementation, I only
considered the case of adding surrounding text, which resulted in an
additional value being added to the '>' mark when removing the
surrounding text. This patch addresses and resolves that issue.