breaking-change: Use `<C-e>` to toggle code block
The `inline_code()` function is now deprecated and should not be used.
Please use the `code()` function instead.

The keybinding for toggling a code block has been changed to `<C-e>`.
This is because many Vim users have remapped `<C-c>` to the Esc key.
Moreover, `<C-e>` is the default keybinding for code blocks on platforms
like Notion, Discord, and GitHub.